+++ djpcms/contrib/admin/templatetags/admin_list.py
@@ -1,1 +1,0 @@
-from django.contrib.admin.templatetags.admin_list import *