+++ djpcms/contrib/admin/templatetags/admin_modify.py
@@ -1,1 +1,0 @@
-from django.contrib.admin.templatetags.admin_modify import *