+++ djpcms/contrib/jdep/static.py
@@ -98,18 +98,38 @@ def vrun(command = ''):
 class ServerInstaller(object):
     
     def config(self, release = True):
-        from fabric.api import env, run
         if release:
             vrun('python server.py')
         else:
             # very problematic to debug this statement. Need a better way.
-            exec(env.server_script)
+            exec(self.env.server_script)
+    
+    def config_files(self, environ, release = True):
+        pass
+    
+    def __get_env(self):
+        from fabric.api import env
+        return env
+    env = property(__get_env)
 
+    def info(self, data):
+        pass
+    
     def install(self, release = True):
         return self
     
     def reboot(self):
         pass
+    
+
+class NginxBase(ServerInstaller):
+    
+    def config_files(self, environ, release = True):
+        rs = [environ.domain_name] + environ.redirects
+        v  = []
+        for r in rs:
+            v.append(r.replace('.','\.'))
+        environ.nginx_redirects = '|'.join(v)
 
 
 class nginx_apache_wsgi(ServerInstaller):
@@ -122,6 +142,7 @@ class nginx_apache_wsgi(ServerInstaller):
         return 'nginx + apache (mod_wsgi)'
     
     def config_files(self, environ, release = True):
+        super(nginx_apache_wsgi,self).config_files(environ,release)
         # Create the config files. Function called from remote server
         environ['apps']  = application_map().values()
         dir = None if not release else environ['confdir']
@@ -166,18 +187,35 @@ class nginx_apache_wsgi(ServerInstaller):
         sudo('/etc/init.d/apache2 restart')
         return self
 
-    def info(self):
-        from fabric.api import env
-        print('apache port:        %(apache_port)s' % env)
+    def info(self, data):
+        data['apache port'] = self.env['redirect_port']
         
 
 
-class twisted_wsgi(ServerInstaller):
-    '''Twisted web wsgi server'''
-     
+class nginx_twisted(NginxBase):
+    '''Twisted behind nginx server'''
+    nginx  = 'nginx.conf'
+    
     def __str__(self):
-        return 'Twisted web wsgi'
+        return 'Twisted behind nginx'
+    
+    def info(self, data):
+        data['twisted port'] = self.env['redirect_port']
+        
+    def config_files(self, environ, dir = None, release = True):
+        super(nginx_twisted,self).config_files(environ,release)
+        # Create the config files. Function called from remote server
+        environ['apps']  = application_map().values()
+        if dir is None:
+            dir = None if not release else environ['confdir']
+        environ['nginx']  = config_file(self.nginx,environ=environ,dir=dir)
+        if not release:
+            from __builtin__ import globals
+            g = globals()
+            g['script_result'] = environ
+    
     
         
+        
 server_types = {'nginx-apache-mod_wsgi':nginx_apache_wsgi(),
-                'twisted':twisted_wsgi}
+                'nginx-twisted':nginx_twisted()}