+++ djpcms/contrib/social/providers/__init__.py
@@ -1,2 +1,8 @@
-import twitter
-import google
+def safeimport(name):
+    try:
+        __import__(name)
+    except ImportError:
+        pass
+
+safeimport('twitter')
+safeimport('google')